# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sierra Sync API
  description: |
    Real-time lead and sales data synchronization platform between Sierra Interactive and popular CRMs.
    
    ## Authentication
    The API uses JWT Bearer tokens for authentication. Obtain a token via the `/auth/login` endpoint.
    
    ## Rate Limiting
    - Free tier: 100 requests/minute
    - Starter: 500 requests/minute
    - Professional: 2000 requests/minute
    - Enterprise: 10000 requests/minute
    
    ## Webhooks
    Register webhooks to receive real-time updates on lead and sync events.
  version: 1.0.0
  contact:
    name: Sierra Sync Support
    email: support@sierrasync.com
    url: https://sierrasync.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.sierrasync.com/v1
    description: Production server
  - url: https://staging-api.sierrasync.com/v1
    description: Staging server
  - url: http://localhost:3001/v1
    description: Local development
security:
  - bearerAuth: []
  - apiKey: []
...
paths:
...
  /organizations:
    get:
      tags: [Organizations]
      summary: List organizations
      operationId: listOrganizations
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Organizations list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationList'

    post:
      tags: [Organizations]
      summary: Create organization
      operationId: createOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationRequest'
      responses:
        '201':
          description: Organization created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
  /organizations/{id}:
    get:
      tags: [Organizations]
      summary: Get organization details
      operationId: getOrganization
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      responses:
        '200':
          description: Organization details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    OrganizationId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        subscriptionTier:
          type: string
          enum: [free, starter, professional, enterprise]
        maxUsers:
          type: integer
        maxIntegrations:
          type: integer
        maxSyncsPerMonth:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    CreateOrganizationRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        subscriptionTier:
          type: string
          enum: [free, starter, professional, enterprise]
...
    OrganizationList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Organization'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...